fix(destination)!: replace ip_port_subscribers with workload_subscribers (#14612)
* fix(destination): drop ip and port metric labels

The destination control's ip_port_subscriber and ip_port_updates metrics are
potentailly exposed for every IP that is discovered. This is high cardinality
and puts undo pressure on the metrics infrastructure.

This change drops these labels, so that there is a single ip_port_subscriber and
ip_port_updates timeseries.

* rename metric to workload_subscribers