fix(destination): fix workload subscribers metric (#14683)
The destination controller has a gauge metric called workload_subscribers which tracks the number of subscribers to an endpoint profile. However, since we have dropped all labels from this metric due to high cardinality, we are now re-using the same gauge for all ip:port publishers. This means that each time we set the gauge value, it overrides the previous value and only tracks the number of subscribers to the most recently subscribed ip:port combination.

We move this logic up so that the workload_subscribers gauge instead tracks the total number of subscribers across all ip:port publishers.

Signed-off-by: Ivan Porta <porta.ivan@outlook.com>