docs(cli): a comment about cli image `ENTRYPOINT`
this image is used within CI, and broadly as a "layer" to place the cli
binaries in a final image.

the cli image itself however, is not currently intended for use directly
via `docker build`, and will consequently yield an error response if
built directly.

this commit adds a comment providing some information about this.

see https://github.com/linkerd/linkerd2/pull/14360 for more.

Signed-off-by: katelyn martin <kate@buoyant.io>