fix(bin): restore `bin/docker-build` functionality (#14698)
this branch makes changes to scripts in the `bin/` directory, to
restore the end-to-end local development workflow described in
`BUILD.md` here:
https://github.com/linkerd/linkerd2/blob/main/BUILD.md#comprehensive

now that we no longer include a `bin-cli` in our published artifacts,
we can remove that image from the list of images that the `_docker.sh`
helper script is aware of. this means we now refrain from (a)
building the image when `bin/docker-build` is run, and (b) loading the
image when `bin/image-load` is run.

some related comments are updated in this branch, while we are here.

---

* refactor: remove `cli-bin` from `_docker.sh` images

this helper script wrapping docker is used by various scripts in the
development workflow.

in https://github.com/linkerd/linkerd2/pull/14360, we removed the
`cli-bin` image from our published artifacts.

Signed-off-by: katelyn martin <kate@buoyant.io>

* nit(bin): update `build-cli-bin` comment

this commit removes mention of docker from this comment. there is no
longer an equivalent `docker-build-cli-bin` script, so this is the way
that we build the cli binary.

Signed-off-by: katelyn martin <kate@buoyant.io>

* fix(bin): remove cli from `bin/docker-build`

this commit removes mention of the cli from our `bin/docker-build`
script.

in https://github.com/linkerd/linkerd2/pull/14360, we removed the cli
image from our published artifacts. moreover, invoking `bin/linkerd`
will call `bin/build-cli-bin`.

so, this commit removes both calls, to the now deleted
`bin/docker-build-cli-bin` script, as well as the `bin/build-cli-bin`
script. this lets the `bin/docker-build` script focus solely on docker
images.

Signed-off-by: katelyn martin <kate@buoyant.io>

* docs(cli): a comment about cli image `ENTRYPOINT`

this image is used within CI, and broadly as a "layer" to place the cli
binaries in a final image.

the cli image itself however, is not currently intended for use directly
via `docker build`, and will consequently yield an error response if
built directly.

this commit adds a comment providing some information about this.

see https://github.com/linkerd/linkerd2/pull/14360 for more.

Signed-off-by: katelyn martin <kate@buoyant.io>

---------

Signed-off-by: katelyn martin <kate@buoyant.io>