chore(test): Make tracing test more robust
Previously, the tracing integration test only checked for a simple existence of a trace with a service name, while not inspecting the actual values inside the trace.

This makes the test check the tags inside a specific trace, both for a process and a span.

In the future, this can be improved even more to inspect all spans in a trace and ensure the relationships between them are correct, but this will require updates to emojivoto and its propagation to be viable.

Signed-off-by: Scott Fleener <scott@buoyant.io>