feat(proxy, proxy-init): Combine proxy and proxy-init and use minimal base image (#14577)
To simplify the docker images we ship, we combine the proxy and proxy-init images into a unified image (named proxy) which includes both the proxy and proxy-init binaries.

To reduce the security surface area of this unified image, we build it on a minimal Wolfi-based runtime image via apko instead of building on `gcr.io/distroless/cc-debian12`. This allows us to avoid including unused packages such as `libssl` which can cause spurious security scan alerts in our images.

In order to build this minimal runtime base image in CI, we start a local temporary registry so that we can push the apko created runtime image and use it as a base image for the proxy image.

We update the Linkerd templates to use this new unified proxy image in the linkerd-init container.

Signed-off-by: Alex Leong <alex@buoyant.io>