test(policy): avoid timeouts due to flakiness
Policy tests are very flaky. Currently one of the main culprits is that
service account creation sometimes isn't caught as an event by the
watcher, blocking `await_service_account` until it times out after 60s.
We already have in place up to 3 retries when calling `cargo nextest`,
but these tests are sequential and the 60s timeouts start accumulating
until we reach the CI job timeout at 20min.

This change lowers the service account creation timeout down to 15s,
understanding that if the watcher catches that event it will do pretty
quickly or else block indefinitely. So better to fail faster and trigger
the test retry ASAP.